Stop printing a refused count as zero, and put an envelope on every record (#21)
A count that prints as 0 because the endpoint carrying it refused is the
same lie as an empty list that means refused, one field down. bili user
printed video_count, total_view and total_like as 0 for a creator with
924 uploads and 17 million followers, and nothing in the output said so.

The five counts on User are now *int64. A nil is a count nobody was told,
which json omits, table and csv render as an empty cell, and a creator
whose totals really are zero still gets a zero. Three outcomes where
there used to be two.

Every record type carries an Envelope: the endpoint that answered, whether
the request was signed, the state the response was sorted into, when, the
size of the body, and a missed map naming the fields that are not there
and what stopped each one. It is tagged out of table and csv and stays in
json, because provenance is worth having on every record and worth a
column on none of them.

video_count was never on x/space/upstat. Issue #13 and the model document
both said it was; upstat carries archive.view and likes and nothing else.
It was a field declared and never written, which is why it printed 0. Its
real source is the listing endpoint's own pagination total, read with
pn=1&ps=1, and bili user is now four requests behind one row.

Fixes the matrix walk while here. It resolved endpoints only from string
literals in the call, so moving the URLs into constants took it from 29
endpoints to 1 with both tests still green. It resolves constants now,
and a fourth test puts a floor under the count.

+++ bili/walk_test.go
@@ -95,7 +95,7 @@ func newFakeGraph() *fakeGraph {
 			"BV_TURING": vid("BV_TURING", 2, "Creator A", "On Computable Numbers"),
 		},
 		users: map[string]*User{
-			"2": {Mid: 2, Name: "Creator A", FollowerCount: 1000},
+			"2": {Mid: 2, Name: "Creator A", FollowerCount: i64(1000)},
 		},
 		related: map[string][]Video{
 			"BV_TURING": {vid("BV_REL1", 3, "Creator B", "Related one"), vid("BV_REL2", 2, "Creator A", "Related two")},
